Move arg parsing to cli module and type it
